PERA-1773 :: Skip info screens after import accounts
